backend/requirements.txt
# FastAPI和Web服务器
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.6
email-validator>=2.0.0

# 数据库
sqlalchemy>=2.0.0
alembic>=1.12.0
pydantic>=2.4.2
pydantic-settings>=2.0.3

# 安全
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
cryptography>=41.0.0

# LLM API客户端
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0
httpx>=0.25.0

# 仿真引擎
numpy>=1.26.0
pandas>=2.1.1
matplotlib>=3.8.0
scipy>=1.11.3

# 测试
pytest>=7.4.2
pytest-asyncio>=0.21.1
httpx>=0.25.0

# 工具
python-dotenv>=1.0.0
tenacity>=8.2.3
loguru>=0.7.2 